requirements.txt
streamlit>=1.35
anthropic>=0.28
chromadb>=0.5
sentence-transformers>=3.0
pymupdf>=1.24
pandas>=2.0
python-docx>=1.1
openpyxl>=3.1
python-dotenv>=1.0
